cleanly separate github clone testing functions
